Added DataStoreModule.kt and refactored LocalPreferences and Token provider to the new datastore module
