+++ internal/datastore/src/main/java/com/hover/stax/internal/datastore/LocalPreferences.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hover.stax.preferences
+package com.hover.stax.internal.datastore
 
 import android.content.Context
