+++ internal/datastore/src/main/java/com/hover/stax/internal/datastore/TokenProvider.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hover.stax.preferences
+package com.hover.stax.internal.datastore
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
