+++ internal/datastore/src/main/java/com/hover/stax/internal/datastore/di/DataStoreModule.kt
@@ -1,0 +1,46 @@
+package com.hover.stax.internal.datastore.di
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.SupervisorJob
+import javax.inject.Singleton
+import androidx.datastore.core.DataStore
+import androidx.datastore.core.handlers.ReplaceFileCorruptionHandler
+import androidx.datastore.preferences.core.PreferenceDataStoreFactory
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.preferencesDataStoreFile
+import androidx.datastore.preferences.SharedPreferencesMigration
+import androidx.datastore.preferences.core.emptyPreferences
+import kotlinx.coroutines.Dispatchers
+
+
+@Module
+@InstallIn(SingletonComponent::class)
+object DataStoreModule {
+
+    @Provides
+    @Singleton
+    fun providesUserPreferencesDataStore(
+        @ApplicationContext context: Context,
+    ): DataStore<Preferences> =
+        PreferenceDataStoreFactory.create(
+            corruptionHandler = ReplaceFileCorruptionHandler(
+                produceNewData = { emptyPreferences() }
+            ),
+            scope = CoroutineScope(Dispatchers.IO + SupervisorJob()),
+            migrations = listOf(
+                SharedPreferencesMigration(
+                    context,
+                    sharedPreferencesName = "stax.datastore"
+                )
+            ),
+        ){
+            context.preferencesDataStoreFile(name = "stax.datastore")
+        }
+}
+
